requirements.txt
flask>=2.0.0
python-dotenv>=1.0.0
together>=0.2.10
sentence-transformers>=2.2.2
fake-useragent>=1.4.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
playwright>=1.40.0
numpy>=1.24.0
scikit-learn>=1.3.0
geopy>=2.4.0
schedule>=1.2.0
requests>=2.25.0
# difflib is part of the Python standard library 